% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pop_fitness}
\alias{check_pop_fitness}
\title{Check if there is anything wrong with pop.fitness calculation during optimization}
\usage{
check_pop_fitness(pop.fitness)
}
\arguments{
\item{pop.fitness}{A list containing the fitness result of each chromosome in a given population based on the fitness function}
}
\value{
A list with error message and whether we need to rerun the fitness calculation
}
\description{
Check if there is anything wrong with pop.fitness calculation during optimization
}
