% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_dis}
\alias{cluster_dis}
\title{Clustering of cell types}
\usage{
cluster_dis(
  count_table,
  cell_cluster_conversion,
  dist_metric = "correlation",
  cluster_metric = "complete",
  top.var.gene = 1000,
  log.transform = T
)
}
\arguments{
\item{count_table}{A matrix containing the expression level of each gene in each cell with gene name as row name and cell name as column name}

\item{cell_cluster_conversion}{A data frame with each row representing information of one cell. First column contains the cell name. Second column contains the corresponding cell type name. Row name of the data frame should be the cell name.}

\item{dist_metric}{A character specifying the metric for distance calculation. Default is "correlation", which uses 1 - Pearson correlation coefficient as distance. Other options are "euclidean" for Euclidean distance and "manhattan" for Manhattan distance.}

\item{cluster_metric}{A character specifying the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete" (default), "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{top.var.gene}{A numeric value specifying the number of top highly variable genes to use for distance calculation. Default is 1000}

\item{log.transform}{A logical value indicating whether to perform log transformation on input count data. Default is TRUE.}
}
\value{
A list containing 3 slots:
\item{distance_matrix}{A \code{dist} object containing the distance between all pairs of cell types.}
\item{hierarchy}{Hierarchical clustering result returned by \code{hclust}.}
\item{dendrogram}{Clustering dendrogram.}
}
\description{
cluster_dis first calculates distance between cell types and then performs hierarchical clustering.
Specifically, expression of cells from the same cell type is averaged to represent the expression level of a cell type.
Then, highly variable genes are selected based on each gene's average expression across cell types.
Expression information of top highly expressed genes is used to calculate distance between cell types.
The calculated distance is finally used for hierarchical clustering.
}
\examples{
data(norm_sc_count)
data(sc_cluster)
cluster_distance=cluster_dis(count_table = norm_sc_count, cell_cluster_conversion = sc_cluster,
                             cluster_metric = "complete", dist_metric = "correlation",
                             log.transform = FALSE, top.var.gene = 1000)
}
