% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_transformation}
\alias{data_transformation}
\title{Data transformation}
\usage{
data_transformation(input_data, trans_type, base = 10)
}
\arguments{
\item{input_data}{A numeric vector of values to transform}

\item{trans_type}{Type of transformation:
\itemize{
\item \code{log}: log transformation
\item \code{square_root}: square root transformation
\item \code{logit}: logit transformation
\item \code{none}: no transformation
}}

\item{base}{Base of log transformation. Default is 10}
}
\value{
A numeric vector of transformed values
}
\description{
data_transformation performs basic data transformations, e.g., log transformation
}
\details{
For log and logit transofmration, if the value before transformation is 0, we will replace it by a small value before the transformation to avoid negative infinity. The small value is calculated as the minimum of all non-zero values from the input divided by 2
}
\examples{
x = sample(1:100, size = 10)
data_transformation(x, trans_type = "log", base=10)
}
