% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{densityscatter}
\alias{densityscatter}
\title{A scatterplot with color corresponding to density}
\usage{
densityscatter(
  x,
  y,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  add.diagonal = F,
  x_low = NULL,
  x_high = NULL,
  y_low = NULL,
  y_high = NULL
)
}
\arguments{
\item{x}{A numeric vector containing the \code{x} coordinates of points in the plot}

\item{y}{A numeric vector containing the \code{y} coordinates of points in the plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{main}{an overall title for the plot}

\item{add.diagonal}{a logical value indicating whether you want to add \code{x = y} line to the plot}

\item{x_low}{A numeric value representing the lower limit of \code{x} coordinate}

\item{x_high}{A numeric value representing the upper limit of \code{x} coordinate}

\item{y_low}{A numeric value representing the lower limit of \code{y} coordinate}

\item{y_high}{A numeric value representing the upper limit of \code{y} coordinate}
}
\description{
Visualize two dimensional data using scatterplot with density information
}
\examples{
require(ggplot2)
x = rnorm(1000, mean = 0, sd = 1)
y = rnorm(1000, mean = 0, sd = 1)
densityscatter(x, y, xlab = "x axis", ylab = "y axis", main = "example plot", add.diagonal = TRUE,
               x_low = 0, x_high = 3, y_low = 0, y_high = 3)
}
