% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_gene_cluster}
\alias{diff_gene_cluster}
\title{Differential gene expression between clusters}
\usage{
diff_gene_cluster(
  pagoda_object,
  cell_cluster_conversion,
  n.core = 1,
  z.threshold = 3
)
}
\arguments{
\item{pagoda_object}{A pagoda2 object}

\item{cell_cluster_conversion}{A data frame with each row representing information of one cell. First column contains the cell name. Second column contains the corresponding cell type name. Row name of the data frame should be the cell name.}

\item{n.core}{Number of cores to use. Default is 1.}

\item{z.threshold}{A numeric value specifying the minimal absolute Z score (adjusted) to report. Default is 3.}
}
\value{
A list containing the differential expression result ("diff_result" slot) and the pagoda2 object ("pagoda.obj" slot)
}
\description{
diff_gene_cluster performs differential gene expression between clusters using Pagoda2
}
\examples{
data(sc_count)
data(sc_cluster)

sc_count.adjust_variance = preprocess_normalize(sc_count, n.core = 2)
diff_expr = diff_gene_cluster(pagoda_object = sc_count.adjust_variance$pagoda.object, cell_cluster_conversion = sc_cluster, n.core = 1)

#Differentially expressed genes for Pvm:
head(diff_expr$diff_result$Pvm)
}
