% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distortion_test}
\alias{distortion_test}
\title{Check distortion of gene expression between scRNA-seq and spatial transcriptomics}
\usage{
distortion_test(prop_vector, group_label)
}
\arguments{
\item{prop_vector}{A numeric vector containing the expression level of one gene from two data modalities}

\item{group_label}{A character vector indicating which value is from which data modality (\code{sc} for scRNA-seq and \code{spatial} for spatial transcriptomics)}
}
\value{
Coefficients from linear and deming regression
}
\description{
distortion_test performs linear regression and deming regression on gene expression from scRNA-seq and spatial transcriptomics to check their agreement
}
\examples{
x = sample(1:10, 100, replace=TRUE)
y = sample(100:1000, 100, replace=TRUE)
label = c(rep("sc", 100), rep("spatial", 100))
distortion_test(c(x,y), label)
}
