% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_AUC_from_combined_pp}
\alias{get_AUC_from_combined_pp}
\title{Calculate AUC per cell type}
\usage{
get_AUC_from_combined_pp(combined.pred.prob, cell_cluster_conversion)
}
\arguments{
\item{combined.pred.prob}{A matrix containing the predicted probability of each cell to each cell type with cell type name as column name.
Each row is one cell and each column is one cell type.}

\item{cell_cluster_conversion}{A data frame with each row representing information of one cell.
First column contains the cell name. Second column contains the corresponding cell type name. Row name of the data frame should be the cell name.}
}
\value{
A numeric vector containing the AUC for each cell type.
}
\description{
get_AUC_from_combined_pp calculates AUC for each cell type given the predicted probability of each cell to each cell type.
}
