% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsFISH_optimize.R
\name{gpsFISH_optimize}
\alias{gpsFISH_optimize}
\title{Genetic algorithm for gene panel selection}
\usage{
gpsFISH_optimize(
  n,
  k,
  OF = fitness,
  popsize = 200,
  keepbest = floor(popsize/10),
  ngen = 50,
  tourneysize = max(ceiling(popsize/10), 2),
  mutprob = 0.01,
  mutfrac = NULL,
  initpop = NULL,
  verbose = 0,
  cluster = NULL,
  save.intermediate = F,
  gene2include.id = NULL,
  gene.weight = NULL,
  earlyterm = ngen,
  converge.cutoff = 0,
  ...
)
}
\arguments{
\item{n}{Total number of genes for gene panel selection.}

\item{k}{Size of a gene panel, i.e., the number of genes in a gene panel.}

\item{OF}{Fitness function, a function that evaluate the fitness of each gene panel. Default is \code{fitness}.
Additional arguments can be passed to \code{OF} through \code{\dots}. See \link[gpsFISH]{fitness} for more details.}

\item{popsize}{Size of the population, i.e., the number of gene panels in a population. Default is 200.}

\item{keepbest}{The \code{keepbest} least fit offspring of each generation are replaced by the \code{keepbest} most fit members of the previous generation. Used to implement elitism. Default is calculated based on \code{popsize}.}

\item{ngen}{The number of generations to run the optimization. Default is 50.}

\item{tourneysize}{The number of individuals involved in each tournament during selection. Default is calculated based on \code{popsize}.}

\item{mutprob}{Mutation probability of each gene in a gene panel. This probability can be set indirectly through \code{mutfrac}. Default is 0.01.}

\item{mutfrac}{The average fraction of offspring that will experience at least one mutation. Only used if mutprob is not supplied.}

\item{initpop}{A numeric matrix specifying the initial population with each row representing one gene panel and each column representing one gene in a gene panel.
The genes are encoded by their location in the column name of \code{full_count_table}.
If not provided, it will be randomly initialized from all the candidate genes.
The final populations from one gene panel selection can be passed as the initial population of the next selection.
Possibly useful if using this function in an adaptive, iterative, or parallel scheme.}

\item{verbose}{An integer controlling the display of progress during optimization.
If \code{verbose} is greater than 0, then the iteration number and best fitness value are displayed at the console every generation.
Otherwise nothing is displayed. Default is zero (no display).}

\item{cluster}{The number of cores to use, i.e. at most how many child processes will be run simultaneously. Must be at least one, and parallelization requires at least two cores.}

\item{save.intermediate}{A logical value specifying whether intermediate populations are outputted to the current working directory. Default is FALSE.}

\item{gene2include.id}{A numeric vector specifying the location of genes in the column name of \code{full_count_table} that must be included in each panel of the population. Default is NULL.}

\item{gene.weight}{A data frame specifying the weight for each gene. Each row is a gene. The first column contains the gene name. The second column contains the gene weight.
Row name of the data frame is gene name.
A gene with higher weight will be (1) more likely to be selected during crossover, (2) less likely to be mutated if it is already in the population, (3) and more likely to be introduced into the population through mutation if it is not in the current population.}

\item{earlyterm}{If \code{earlyterm} is provided, then the optimization will stop when it reaches \code{ngen} generations, or the maximum fitness improvement (fitness difference between two consecutive generations) in the past \code{earlyterm} generations are smaller or equal to \code{converge.cutoff}, whichever comes first. Default is \code{ngen}.}

\item{converge.cutoff}{A numeric value specifying the cutoff of maximum fitness improvement.}

\item{...}{Additional arguments passed to \code{OF}.}
}
\value{
A list with the following elements:
\item{old}{A list holding information about the search progress. Its elements are:}
\item{old$best}{A matrix containing the best gene panel for each generation.}
\item{old$obj}{A numeric vector containing the fitness values corresponding to the gene panels
in \code{old$best}.}
\item{old$avg}{A numeric vector containing the average population fitness value for each generation.}
\item{pop}{A matrix containing the final population, row-sorted
in order of increasing objective function. Each row is an index vector representing
one gene panel.}
\item{obj}{The fitness values corresponding to each row of \code{pop}.}
\item{bestgeneration}{Number of the best generation.}
\item{bestsol}{A vector of length \code{k} containing the best solution. This is the optimized gene panel.}
\item{bestobj}{The fitness value for the best solution.}
\item{diversity}{A numeric vector containing the diversity of the population in each generation.}
\item{ave_confusionMatrix}{Average confusion matrix over all gene panels for each generation.}
\item{best_confusionMatrix}{Confusion matrix of the best gene panel for each generation.}
\item{norm_ave_confusionMatrix}{Average normalized confusion matrix over all gene panels for each generation.}
\item{norm_best_confusionMatrix}{Normalized confusion matrix of the best gene panel for each generation.}
\item{stats_byclass}{Average classification statistics over all gene panels for each generation.}
\item{best_stats_byclass}{Classification statistics of the best gene panel for each generation.}
\item{best_pred_prob}{Prediction probability of each cell in each cell type in each cross validation of the best gene panel for each generation.}
}
\description{
gpsFISH_optimize implements a genetic algorithm for gene panel selection. The function searches for a subset
of a fixed size, \code{k} gene, from the \code{n} candidate gene, such that user-supplied function \code{OF} is minimized at that subset.
The selection step is done by tournament selection based on ranks, and elitism may be used to retain the best solutions from one generation to the next.
The crossover step is done by uniform crossover.
Population objective function values can be evaluated in parallel.
}
\details{
\itemize{
\item The selection strategy we used is tournaments. Specifically, randomly selected candidate gene panels face each other 1 vs. 1.
The one with a higher fitness value will be used as parent. In addition, candidate gene panels with higher fitness values will be more likely to be selected in the tournaments.
\item After having the parent gene panels, uniform crossover is performed to generate the offspring gene panels.
Duplicated genes after uniform crossover will be replaced by randomly sampled genes in the parent candidate gene panels but not in the offspring gene panel.
}
}
