% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_solution}
\alias{initialize_solution}
\title{Initialize a gene panel within a population for genetic algorithm based on differential gene expression}
\usage{
initialize_solution(
  sol.num,
  panel.size,
  diff_expr_result,
  diff_metric,
  diff_metric_cutoff,
  cluster.list,
  gene.list,
  gene2include = NULL
)
}
\arguments{
\item{sol.num}{A numeric value, can be any number.}

\item{panel.size}{Size of a gene panel, i.e., the number of genes in a gene panel}

\item{diff_expr_result}{A list containing the differential expression result. Each slot in the list contains a data frame corresponding to differentially expressed genes in one cell type.
Within each data frame, each row is one gene. Columns contain differential expression statistics of a gene, e.g., "AUC", "z-score", "precision", etc.}

\item{diff_metric}{A character specifying the metric used to select significant differentially expressed genes. Needs to be one of the column names of the data frame in \code{diff_expr_result}, e.g., "AUC".}

\item{diff_metric_cutoff}{A numeric value representing the significance cutoff for differential expression.}

\item{cluster.list}{A character vector of cell type names. It must be the same or a subset of \code{names(diff_expr_result)}}

\item{gene.list}{A character vector of all candidate genes used to initialize the population}

\item{gene2include}{A character vector of genes that must be included in each panel of the population. Default is NULL}
}
\value{
A numeric vector representing an initialized solution with \code{panel.size} genes.
The genes are encoded by their location in \code{gene.list}
}
\description{
initialize_solution initializes one gene panel within a population using information from differential expression analysis.
Specifically, given the panel size and number of cell types, we first calculate on average how many significantly differentially expressed genes we can include from each cell type.
Then we select this number of genes from the differentially expressed genes of each cell type (or all the differentially expressed genes if the number of them is smaller than this number).
We select the most significant genes first followed by less significant ones.
After doing this for all the cell types, for the rest of the panel, we randomly select genes from the rest of the differentially expressed genes.
}
