% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaxCol}
\alias{MaxCol}
\title{Find the index of the maximum value of each row in a matrix.}
\usage{
MaxCol(m)
}
\arguments{
\item{m}{Numeric matrix}
}
\value{
A numeric vector containing the index for each row.
}
\description{
Find the index of the maximum value of each row in a matrix.
}
