% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maxCol_col}
\alias{maxCol_col}
\title{Find the index of the maximum of each row of a matrix}
\usage{
maxCol_col(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
A numeric vector of index of the maximum value for each row of a matrix
}
\description{
Find the index of the maximum of each row of a matrix
}
