% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_column}
\alias{merge_column}
\title{Merge expression of cells from the same cell type}
\usage{
merge_column(count_table, cell_cluster_conversion, current_label)
}
\arguments{
\item{count_table}{A matrix containing the expression level of each gene in each cell with gene name as row name and cell name as column name}

\item{cell_cluster_conversion}{A data frame with each row representing information of one cell. First column contains the cell name. Second column contains the corresponding cell type name. Row name of the data frame should be the cell name.}

\item{current_label}{A character specifying the current cell type name}
}
\value{
A numeric vector with average gene expression for the current cell type
}
\description{
merge_column calculates expression of genes in a cell type by taking the average from cells of the same cell type
}
