% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot confusion matrix}
\usage{
plot_confusion_matrix(confusion.matrix)
}
\arguments{
\item{confusion.matrix}{Confusion matrix returned by \code{gpsFISH_optimize}.}
}
\value{
ggplot2 object
}
\description{
Plot confusion matrix
}
