% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_norm_confusion_matrix}
\alias{plot_norm_confusion_matrix}
\title{Plot normalized confusion matrix}
\usage{
plot_norm_confusion_matrix(confusion.matrix)
}
\arguments{
\item{confusion.matrix}{A normalized confusion matrix returned by \code{gpsFISH_optimize}.}
}
\value{
ggplot2 object
}
\description{
Plot normalized confusion matrix
}
