% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_norm_confusion_matrix_with_dendrogram}
\alias{plot_norm_confusion_matrix_with_dendrogram}
\title{Plot confusion matrix with cell type cluster dendrogram}
\usage{
plot_norm_confusion_matrix_with_dendrogram(confusion.matrix, cluster.distance)
}
\arguments{
\item{confusion.matrix}{A normalized confusion matrix returned by \code{gpsFISH_optimize}.}

\item{cluster.distance}{Object returned by \code{cluster_distance}.}
}
\value{
ggplot2 object
}
\description{
Plot confusion matrix with cell type cluster dendrogram
}
