% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{popDiv}
\alias{popDiv}
\title{Calculate diversity of a population}
\usage{
popDiv(x)
}
\arguments{
\item{x}{A numeric matrix with each row representing one gene panel and each column represent one gene in a gene panel}
}
\value{
A numeric value representing the diversity
}
\description{
Calculate diversity of a population
}
\examples{
pop = matrix(sample(1:1000000, 10000), 100, 100)
popDiv(pop)
}
