% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsFISH_optimize.R
\name{predict_MNB}
\alias{predict_MNB}
\title{Predict method for multinomial naive bayes model}
\usage{
predict_MNB(object, newdata = NULL, type = c("class", "prob"))
}
\arguments{
\item{object}{Object of class inheriting from "\code{multinomial_naive_bayes}" from the \code{naivebayes} package.}

\item{newdata}{Matrix with non-negative integer predictors (only numeric matrix is accepted).}

\item{type}{if "class", new data points are classified according to the highest posterior probabilities.
If "prob", the posterior probabilities for each class are returned.}
}
\value{
A factor with class labels corresponding to the maximal conditional posterior probabilities (if \code{type = class})
or a matrix with class label specific conditional posterior probabilities (if \code{type = prob}) .
}
\description{
Classification based on the Multinomial Naive Bayes model.
}
