% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_normalize}
\alias{preprocess_normalize}
\title{Adjust variance of scRNA-seq expression data using Pagoda2}
\usage{
preprocess_normalize(count_table, n.core = 1)
}
\arguments{
\item{count_table}{A matrix containing the expression level of each gene in each cell with gene name as row name and cell name as column name}

\item{n.core}{Number of cores to use. Default is 1.}
}
\value{
A list with:
\item{sparse.matrix}{the adjusted count matrix}
\item{pagoda.object}{the pagoda2 object}
}
\description{
preprocess_normalize takes scRNA-seq expression data as input and adjust the variance to normalize the extent to which genes with different expression magnitudes will contribute to the downstream anlaysis
}
\examples{
data(sc_count)
preprocess_normalize(sc_count, n.core = 2)
}
