% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probe_count}
\alias{probe_count}
\title{Probe count per gene.}
\format{
A data frame with 27308 rows and 3 columns. Each row corresponds to one gene. Of note, one gene could appear in multiple rows. For the 3 columns:
\describe{
\item{gene_name}{Gene name.}
\item{chr}{The chromosome that the gene is located.}
\item{probe_count}{Number of probe count.}
}
}
\usage{
probe_count
}
\description{
A data frame containing the number of probe count that can be designed to target a gene
}
\keyword{datasets}
