% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relative_freq}
\alias{relative_freq}
\title{Calculate relative expression of genes in a given cell type}
\usage{
relative_freq(count_matrix, gene_list, cluster_label, cell_cluster_conversion)
}
\arguments{
\item{count_matrix}{A matrix containing the expression level of each gene in each cell with gene name as row name and cell name as column name}

\item{gene_list}{A character vector of gene names for which we want to calculate the relative expression. The gene name should match the row name of \code{count_matrix}}

\item{cluster_label}{A character vector with one element representing the cell type name}

\item{cell_cluster_conversion}{A data frame with each row representing information of one cell. First column contains the cell name. Second column contains the corresponding cell type name. Row name of the data frame should be the cell name.}
}
\value{
A numeric vector of relative expression of each gene in the cell type specified by cluster_label
}
\description{
relative_freq calculates relative expression of genes in a given cell type
}
\examples{
data(spatial_count)
data(spatial_cluster)
cell.type = unique(spatial_cluster$class_label)[1]
rf = relative_freq(spatial_count, rownames(spatial_count), cell.type, spatial_cluster)
rf
}
