% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep.row}
\alias{rep.row}
\title{Replicate a vector row-wise to generate a matrix}
\usage{
\method{rep}{row}(x, n)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{Number of times to repeat \code{x}.}
}
\value{
A numeric matrix.
}
\description{
Replicate a vector row-wise to generate a matrix
}
\examples{
x = c(1:10)
rep.row(x, 3)
}
