% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roc_cal}
\alias{roc_cal}
\title{Calculate AUC for one cell type from the predicted probability of each cell to each cell type.}
\usage{
roc_cal(x, pred.prob)
}
\arguments{
\item{x}{A numeric value indicating the column of \code{pred.prob}, i.e., the cell type to calculate AUC.}

\item{pred.prob}{A matrix containing the predicted probability of each cell to each cell type with cell type name as column name.
Each row is one cell and each column is one cell type.}
}
\value{
AUC of the current cell type.
}
\description{
Calculate AUC for one cell type from the predicted probability of each cell to each cell type.
}
