% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc2spatial.R
\name{sc2spatial}
\alias{sc2spatial}
\title{Title}
\usage{
sc2spatial(
  count_table,
  cell_cluster_conversion = NULL,
  relative_prop = NULL,
  sample_new_levels = NULL,
  use_average_cluster_profiles = NULL,
  rate = 1,
  cluster_size_max = 1e+24,
  cluster_size_min = 0,
  sampling_type,
  nCV = NULL,
  simulation_parameter = NULL,
  simulation_model = NULL
)
}
\arguments{
\item{simulation_model}{}
}
\value{

}
\description{
Title
}
