% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_cluster}
\alias{sc_cluster}
\title{scRNA-seq count data cell annotation.}
\format{
A data frame with 550 rows and 2 columns:
\describe{
\item{cell_name}{name of each cell}
\item{class_label}{annotated cell type of each cell}
...
}
}
\usage{
sc_cluster
}
\description{
A dataset containing the cell type label of the 550 cells in the scRNA-seq data.
}
\keyword{datasets}
