% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_count}
\alias{sc_count}
\title{scRNA-seq count data.}
\format{
A count matrix with 19,972 rows and 550 columns:
}
\usage{
sc_count
}
\description{
A dataset containing the expression of 19,972 genes in 550 cells.
}
\keyword{datasets}
