% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Seurat_clustering}
\alias{Seurat_clustering}
\title{UMAP plot using Seurat}
\usage{
Seurat_clustering(count_table, cell_cluster_conversion)
}
\arguments{
\item{count_table}{Expression matrix with each row representing a gene and each column representing a cell.}

\item{cell_cluster_conversion}{A data frame with each row representing information of one cell.
First column contains the cell name. Second column contains the corresponding cell type name. Row name of the data frame should be the cell name.}
}
\value{
ggplot2 object
}
\description{
UMAP plot using Seurat
}
