% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_params}
\alias{simulation_params}
\title{Pre-trained Bayesian model based on the Codeluppi dataset.}
\format{
A list with 7 slots:
\describe{
\item{data2fit}{A data frame containing the input data for fitting the Bayesian model.}
\item{model.summary}{A named list with elements \code{summary} and \code{c_summary}, which contain summaries of a stanfit object. To reduce the size of simulation_params, we removed information stored in this slot since it is not used in downstream examples, e.g., simulate spatial transcriptomics data.}
\item{fixed.effect}{A named list with elements \code{summary} and \code{c_summary}, which contain summaries of specified parameters.}
\item{c_i_full}{A matrix containing the estimated gene specific intercept \eqn{c_i} for each gene.}
\item{gamma_i_full}{A matrix containing the estimated gene specific coefficient \eqn{\gamma_i} for each gene.}
\item{lib.size}{A numeric vector containing the library size, i.e., total number of molecule per cell, of cells in the spatial transcriptomics data.}
\item{posterior}{A list containing the extracted samples of variables in the Bayesian model from their posterior distribution.}
}
}
\usage{
simulation_params
}
\description{
A list object returned by the simulation_training_ZINB function and then trimmed by the simulation_training_ZINB_trim function. The model is trained on the Codeluppi dataset.
}
\keyword{datasets}
