% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_training_ZINB.R
\name{simulation_training_ZINB}
\alias{simulation_training_ZINB}
\title{Platform effect estimation using Bayesian modelling}
\usage{
simulation_training_ZINB(
  sc_count,
  spatial_count,
  overlap_gene,
  unique_cluster_label,
  sc_cluster,
  spatial_cluster,
  outputpath,
  optimizer,
  mcmc.check = FALSE,
  saveplot = FALSE,
  num.iter = 2000,
  num.chain = 4,
  num.core = 1,
  max.treedepth = 10,
  seed = 3
)
}
\arguments{
\item{sc_count}{A count matrix containing the expression of each gene in each cell from the scRNA-seq data. Each row represents one gene and each column represents one column, with gene name as row name and cell name as column name.}

\item{spatial_count}{A count matrix containing the expression of each gene in each cell from the spatial transcriptomics data. Each row represents one gene and each column represents one column, with gene name as row name and cell name as column name.}

\item{overlap_gene}{A character vector of gene names that show up in both scRNA-seq and spatial transcriptomics data}

\item{unique_cluster_label}{A character vector of cell type names that show up in both scRNA-seq and spatial transcriptomics data}

\item{sc_cluster}{A data frame with two columns and cell name as row name. The first column ("cell_name") contains the name of cells from scRNA-seq data and the second column ("class_label") contains each cell's corresponding cell type.}

\item{spatial_cluster}{A data frame with two columns and cell name as row name. The first column ("cell_name") contains the name of cells from spatial transcriptomics data and the second column ("class_label") contains each cell's corresponding cell type.}

\item{outputpath}{A character specifying the path for output}

\item{optimizer}{Optimizer for Bayesian model fitting.
\itemize{
\item \code{variational_inference}: The Bayesian model will be fitted using the \code{meanfield} variational inference
\item \code{sampling}: The Bayesian model will be fitted using the No-U-Turn sampler variant of Hamiltonian Monte Carlo
}}

\item{mcmc.check}{a logical value indicating whether to do MCMC diagnostics. Default is FALSE. Keep it as FALSE if \code{optimizer} is \code{variational_inference}}

\item{saveplot}{a logical value indicating whether to save plot. Default is FALSE.}

\item{num.iter}{A positive integer specifying the number of iterations for each chain (including warmup). The default is 2000. It is recommended to use at least 10000 if \code{optimizer} is \code{variational_inference}}

\item{num.chain}{A positive integer specifying the number of Markov chains. The default is 4.}

\item{num.core}{The number of cores to use when executing the Markov chains in parallel. The default is to 1 core. However, we recommend setting it to be as many processors as the hardware and RAM allow (up to the number of chains).}

\item{max.treedepth}{A positive integer specifying the maximum tree depth. The default is 10.}

\item{seed}{The seed for random number generation, defaulting to 3.}
}
\value{
A list with the following elements:
\item{distortion_model}{A stanfit object containing the fitted Bayesian model.}
\item{data2fit}{A data frame containing the input data for fitting the Bayesian model:}
\itemize{
\item \code{sp.count}: number of molecule per gene per cell from the spatial transcriptomis data
\item \code{sc.prop}: relative expression per gene per cell from the scRNA-seq data
\item \code{gene}: gene name
\item \code{cell}: cell name
\item \code{cluster}: cell type name of cells in the spatial transcriptomics data
\item \code{cs}: cell size, i.e., total number of molecule per cell, of cells in the spatial transcriptomics data
}

\item{model.summary}{A named list with elements \code{summary} and \code{c_summary}, which contain summaries of a stanfit object.}
\item{fixed.effect}{A named list with elements \code{summary} and \code{c_summary}, which contain summaries of specified variables.}
\item{c_i_full}{A matrix containing the estimated gene specific intercept \eqn{c_i} for each gene.}
\item{gamma_i_full}{A matrix containing the estimated gene specific coefficient \eqn{\gamma_i} for each gene.}
\item{lib.size}{A numeric vector containing the library size, i.e., total number of molecule per cell, of cells in the spatial transcriptomics data.}
}
\description{
simulation_training_ZINB estimates platform effects between scRNA-seq and spatial transcriptomics technologies using Bayesian modelling. The Bayesian model is fitted using rstan
}
\examples{
data(sc_count)
data(sc_cluster)
data(spatial_count)
data(spatial_cluster)
overlap_gene = intersect(unique(rownames(sc_count)),unique(rownames(spatial_count)))
unique_cluster_label=intersect(as.character(unique(sc_cluster$class_label)),
                               as.character(unique(spatial_cluster$class_label)))
outputpath = "~"
simulation.params=simulation_training_ZINB(sc_count = sc_count,
                                           spatial_count = spatial_count,
                                           overlap_gene = overlap_gene,
                                           unique_cluster_label = unique_cluster_label,
                                           sc_cluster = sc_cluster,
                                           spatial_cluster = spatial_cluster,
                                           outputpath = outputpath,
                                           optimizer = "variational_inference",
                                           mcmc.check = FALSE,
                                           num.iter = 300,
                                           num.chain = 4,
                                           num.core = 4,
                                           max.treedepth = 10,
                                           seed = 3,
                                           saveplot = FALSE)

}
\seealso{
\code{\link[rstan]{stan}},  \code{\link[rstan]{vb}} from the \code{rstan} package
}
