% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc2spatial.R
\name{simulation_ZINB}
\alias{simulation_ZINB}
\title{Simulating spatial transcriptomics data from scRNA-seq data using Bayesian modelling}
\usage{
simulation_ZINB(
  gene_list,
  cell_list,
  cell_cluster_conversion,
  relative_prop,
  simulation_parameter,
  sample_new_levels,
  use_average_cluster_profiles = FALSE
)
}
\arguments{
\item{gene_list}{A numeric vector containing the location of genes we want to simulate among all genes in the scRNA-seq data.}

\item{cell_list}{A numeric vector containing the location of cells we want to simulate among all cells in the scRNA-seq data.}

\item{cell_cluster_conversion}{A character vector containing the cell type of each cell.}

\item{relative_prop}{A list with two elements:
\itemize{
\item "cluster.average": A matrix containing the relative expression of each gene in each cell type with gene name as row name and cell type name as column name.
The denominator for relative expression calculation needs to be all genes in the transcriptome before filtering out lowly expressed genes.
\item "cell.level": A matrix containing the relative expression of each gene in each cell with gene name as row name and cell name as column name.
The denominator for relative expression calculation needs to be all genes in the transcriptome before filtering out lowly expressed genes.
}}

\item{simulation_parameter}{A simulation model returned by simulation_training_ZINB_trim.}

\item{sample_new_levels}{A character specifying how simulation is performed for genes we have observed in the data used to train the Bayesian model.
Specifically, during the training of the Bayesian model, we have estimations of the platform effect for genes in the training data.
When we simulate spatial transcriptomics data for genes we have already seen in the training data,
we can use their estimated platform effect \eqn{\gamma_i} and \eqn{c_i} ("old_levels"), or we can randomly sample \eqn{\gamma_i} and \eqn{c_i} from their posterior distribution ("random").
For genes not in the training data, we will randomly sample \eqn{\gamma_i} and \eqn{c_i} from their posterior distribution since we don't have their estimation.}

\item{use_average_cluster_profiles}{A logical value indicating if we want to use relative expression per cell type as the relative expression input for simulating spatial transcriptomics data.
If TRUE, then value in \code{relative_prop$cluster.average} is used. And for each gene, cells from the same cell type will have the same value.
If FALSE, then we use the relative expression per cell as input, i.e., \code{relative_prop$cell.level}.
Default is FALSE.}
}
\value{
A matrix containing the simulated spatial transcriptomics data. Rows and columns are the same with \code{count_table}.
}
\description{
Simulating spatial transcriptomics data from scRNA-seq data using Bayesian modelling
}
