% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_cluster}
\alias{spatial_cluster}
\title{Spatial transcriptomics count data cell annotation.}
\format{
A data frame with 537 rows and 2 columns:
\describe{
\item{cell_name}{name of each cell}
\item{class_label}{annotated cell type of each cell}
...
}
}
\usage{
spatial_cluster
}
\description{
A dataset containing the cell type label of the 537 cells in the spatial transcriptomics data.
}
\keyword{datasets}
