% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_count}
\alias{spatial_count}
\title{Spatial transcriptomics count data.}
\format{
A count matrix with 33 rows and 537 columns:
}
\usage{
spatial_count
}
\description{
A dataset containing the expression of 33 genes in 537 cells.
}
\keyword{datasets}
