% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zinb_generator}
\alias{zinb_generator}
\title{Generate random values from a zero-inflated negative binomial distribution.}
\usage{
zinb_generator(n, size, pie, mu)
}
\arguments{
\item{n}{Number of random values to return.}

\item{size}{The dispersion paramater used in ‘dnbinom’.}

\item{pie}{The zero-inflation parameter.}

\item{mu}{Alternative parametrization of negative binomial distribution via mean. The mean parameter used in ‘dnbinom’.}
}
\value{
A numeric vector containing the randomly generated values.
}
\description{
Generate random values from a zero-inflated negative binomial distribution.
}
\examples{
zinb_generator(30, size = rep(2, 30), pie = 0.1, mu = rep(5, 30))
}
