% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc2spatial.R
\name{ZINB_predict}
\alias{ZINB_predict}
\title{Simulating spatial transcriptomics data from scRNA-seq data using Bayesian modelling}
\usage{
ZINB_predict(
  sc_prop,
  simulation_parameter,
  sample_new_levels,
  gene_list,
  cell_list,
  num_gene,
  num_cell
)
}
\arguments{
\item{sc_prop}{A matrix containing the relative expression for each gene in each cell.}

\item{simulation_parameter}{A simulation model returned by simulation_training_ZINB_trim.}

\item{sample_new_levels}{A character specifying how simulation is performed for genes we have observed in the data used to train the Bayesian model.
Specifically, during the training of the Bayesian model, we have estimations of the platform effect for genes in the training data.
When we simulate spatial transcriptomics data for genes we have already seen in the training data,
we can use their estimated platform effect \eqn{\gamma_i} and \eqn{c_i} ("old_levels"), or we can randomly sample \eqn{\gamma_i} and \eqn{c_i} from their posterior distribution ("random").
For genes not in the training data, we will randomly sample \eqn{\gamma_i} and \eqn{c_i} from their posterior distribution since we don't have their estimation.}

\item{gene_list}{A character vector containing the name of genes we want to simulate.}

\item{cell_list}{A character vector containing the name of cells we want to simulate.}

\item{num_gene}{Number of genes.}

\item{num_cell}{Number of cells.}
}
\value{
A list with elements:
\item{simu_count_matrix}{A matrix containing the simulated spatial transcriptomics data.}
\item{alpha.global}{Randomly sampled \eqn{\alpha} from its posterior distribution.}
\item{beta.global}{Randomly sampled \eqn{\beta} from its posterior distribution.}
\item{zi.global}{Randomly sampled \eqn{\pi} from its posterior distribution.}
\item{mu_gamma}{Randomly sampled \eqn{\mu_\gamma} from its posterior distribution.}
\item{sigma_gamma}{Randomly sampled \eqn{\sigma_\gamma} from its posterior distribution.}
\item{mu_c}{Randomly sampled \eqn{\mu_c} from its posterior distribution.}
\item{sigma_c}{Randomly sampled \eqn{\signa_c} from its posterior distribution.}
\item{gamma_i_gene}{Simulated \eqn{\gamma_i} for each gene.}
\item{c_i_gene}{Simulated \eqn{c_i} for each gene.}
\item{new.cell.size}{Simulated cell depth, i.e., total number of molecules per cell.}
}
\description{
Simulating spatial transcriptomics data from scRNA-seq data using Bayesian modelling
}
