% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_normalize}
\alias{preprocess_normalize}
\title{Adjust variance of count data using Pagoda2}
\usage{
preprocess_normalize(count_table, n.core = 1)
}
\arguments{
\item{count_table}{A matrix containing the expression level of each gene in each cell with gene name as row name and cell name as column name}

\item{n.core}{Number of cores to use. Default is 1.}
}
\value{
A list containing the adjusted count matrix ("sparse.matrix" slot) and the Pagoda2 object ("pagoda.object" slot)
}
\description{
preprocess_normalize takes count data as input and adjust the variance to normalize the extent to which genes with different expression magnitudes will contribute to the downstream anlaysis
}
\examples{
data(sc_count)
preprocess_normalize(sc_count, ncore = 2)
}
