% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsample_sc}
\alias{subsample_sc}
\title{Subsample scRNA-seq data}
\usage{
subsample_sc(
  cell_list,
  cell_cluster_conversion,
  rate = 1,
  cluster_size_max = 1000,
  cluster_size_min = 1,
  sampling_type = "Subsampling_by_cluster",
  nCV = 1
)
}
\arguments{
\item{cell_list}{A character vector containing the cell names to subsample from.}

\item{cell_cluster_conversion}{A data frame with each row representing information of one cell.
First column contains the cell name. Second column contains the corresponding cell type name. Row name of the data frame should be the cell name.
Only needed if \code{sampling_type} is "Subsampling_by_cluster".}

\item{rate}{A value between 0 and 1 specifying the proportion of cells we want to keep for each cell type during subsampling. 0.8 means we keep 80\% of cells for each cell type. Default is 1.}

\item{cluster_size_max}{Maximum number of cells to keep for each cell type during subsampling. Default is 1000. Only needed if \code{sampling_type} is "Subsampling_by_cluster".}

\item{cluster_size_min}{Minimum number of cells to keep for each cell type during subsampling. Default is 1. Only needed if \code{sampling_type} is "Subsampling_by_cluster".}

\item{sampling_type}{A character specifying how the subsample is performed.
"Subsampling_by_cluster" means subsampling cells from each cell type separately. "Subsampling" means subsampling all cells together by mixing cells from all cell types.
Default is "Subsampling_by_cluster".}

\item{nCV}{Number of cross validation to perform on the subsampled data. This is to ensure that we have enough cells for cross validation. Default is 1. Only needed if \code{sampling_type} is "Subsampling_by_cluster".}
}
\value{
A list with elements:
\item{cell_loc}{Location of selected cells in \code{cell_list} after subsampling.}
\item{cell_class}{Cell type of selected cells after subsampling.}
}
\description{
subsample_sc takes cell name of cells in a scRNA-seq data as input and subsample it by selecting a subset of cells.
}
\examples{
data(sc_count)
data(sc_cluster)

#number of cells per cell type
table(sc_cluster[colnames(sc_count), "class_label"])

#subsample by cell type
subsample_sc_info = subsample_sc(cell_list = colnames(sc_count),
                                 cell_cluster_conversion = sc_cluster,
                                 rate = 1,
                                 cluster_size_max = 30,
                                 cluster_size_min = 5,
                                 sampling_type = "Subsampling_by_cluster",
                                 nCV = 5)

#number of cells per cell type after subsample
table(subsample_sc_info$cell_class)

#subsample cells from all cell types together
dim(sc_count)
subsample_sc_info = subsample_sc(cell_list = colnames(sc_count),
                                  rate = 0.5,
                                  cell_cluster_conversion = sc_cluster,
                                  sampling_type = "Subsampling")
length(subsample_sc_info$cell_loc)
table(subsample_sc_info$cell_class)
}
