% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsampling_all}
\alias{subsampling_all}
\title{Subsample scRNA-seq data by mixing cells from all cell types}
\usage{
subsampling_all(count_table, rate = 1)
}
\arguments{
\item{count_table}{A matrix containing the expression level of each gene in each cell with gene name as row name and cell name as column name.}

\item{rate}{A value between 0 and 1 specifying the proportion of cells we want to keep for each cell type during subsampling. 0.8 means we keep 80\% of cells for each cell type. Default is 1.}
}
\value{
A subsampled matrix
}
\description{
Subsample scRNA-seq data by mixing cells from all cell types
}
\examples{
data(sc_count)
dim(sc_count)
subsample_sc_count = subsampling_all(sc_count, 0.5)
dim(subsample_sc_count)
}
