% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsampling_by_cluster}
\alias{subsampling_by_cluster}
\title{Subsample scRNA-seq data by cell type}
\usage{
subsampling_by_cluster(
  cell_list,
  cell_cluster_conversion,
  rate = 1,
  cluster_size_max = 1000,
  cluster_size_min = 1,
  nCV = 1
)
}
\arguments{
\item{cell_list}{A character vector containing the cell names to subsample from.}

\item{cell_cluster_conversion}{A data frame with each row representing information of one cell.
First column contains the cell name. Second column contains the corresponding cell type name. Row name of the data frame should be the cell name.}

\item{rate}{A value between 0 and 1 specifying the proportion of cells we want to keep for each cell type during subsampling. 0.8 means we keep 80\% of cells for each cell type. Default is 1.}

\item{cluster_size_max}{Maximum number of cells to keep for each cell type during subsampling. Default is 1000.}

\item{cluster_size_min}{Minimum number of cells to keep for each cell type during subsampling. Default is 1.}

\item{nCV}{Number of cross validation to perform on the subsampled data. This is to ensure that we have enough cells for cross validation. Default is 1.}
}
\value{
A list with elements:
\item{cell_loc}{Location of selected cells in \code{cell_list} after subsampling.}
\item{cell_class}{Cell type of selected cells after subsampling.}
}
\description{
Subsample scRNA-seq data by cell type
}
